Refactor to centralise ScalePresentationArea, so things are scale from the MainMenu onwards.
